+++ ChangeLog.md
@@ -12,6 +12,10 @@ Behavior changes:
 
 Other enhancements:
 
+* Add flag `--[no-]semaphore` (default: disabled) to Stack's `build` command,
+  to allow GHC to use a system semaphore to perform compilation in parallel when
+  possible. Supported, by default, by GHC 9.10.1 or later.
+
 Bug fixes:
 
 ## v3.9.3 - 2026-02-19